ml-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
scikit-learn==1.5.0
pandas==2.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
